searcher/requirements.txt
fastapi
uvicorn
pydantic
psycopg2-binary
numpy
requests